+++ src/main/java/com/kr/lg/module/login/model/dto/GoogleLoginRequestDto.java
@@ -1,9 +1,10 @@
-package com.kr.lg.module.login.model.kakao;
+package com.kr.lg.module.login.model.dto;
 
+import com.kr.lg.module.login.model.google.GoogleProp;
 import lombok.Data;
 
 @Data
-public class KakaoLoginRequestDto {
+public class GoogleLoginRequestDto {
 
     private String clientId;    // 애플리케이션의 클라이언트 ID
     private String redirectUri; // Google 로그인 후 redirect 위치
@@ -18,11 +19,11 @@ public class KakaoLoginRequestDto {
     private String loginHint;   // 애플리케이션이 인증하려는 사용자를 알고 있는 경우 이 매개변수를 사용하여 Google 인증 서버에 힌트를 제공
     private String prompt;  // default: 처음으로 액세스를 요청할 때만 사용자에게 메시지가 표시
 
-    public KakaoLoginRequestDto(KakaoProp kakaoProp, String authCode) {
-        this.clientId = kakaoProp.getKakaoClientId();
-        this.clientSecret = kakaoProp.getKakaoSecret();
+    public GoogleLoginRequestDto(GoogleProp googleProp, String authCode) {
+        this.clientId = googleProp.getGoogleClientId();
+        this.clientSecret = googleProp.getGoogleSecret();
         this.code = authCode;
-        this.redirectUri = kakaoProp.getKakaoRedirectLoginUri();
+        this.redirectUri = googleProp.getGoogleRedirectUri();
         this.grantType = "authorization_code";
     }
 }
