--- src/main/java/com/kr/lg/module/login/service/GoogleService.java
@@ -1,9 +1,0 @@
-package com.kr.lg.module.login.service;
-
-import com.kr.lg.module.login.model.google.GoogleLoginDto;
-import com.kr.lg.module.login.model.google.GoogleLoginRequestDto;
-
-public interface GoogleService {
-
-    GoogleLoginDto googleOAuth(GoogleLoginRequestDto requestDto) throws Exception;
-}
