+++ src/main/java/com/kr/lg/module/login/service/LoginService.java
@@ -1,0 +1,7 @@
+package com.kr.lg.module.login.service;
+
+import com.kr.lg.db.entities.UserTb;
+
+public interface LoginService {
+    String redirect(UserTb userTb) ;
+}
