+++ src/main/java/com/kr/lg/module/user/service/UserService.java
@@ -2,6 +2,7 @@ package com.kr.lg.module.user.service;
 
 import com.kr.lg.db.entities.UserTb;
 import com.kr.lg.exception.LgException;
+import com.kr.lg.module.login.model.dto.LoginDto;
 import com.kr.lg.module.user.model.req.EnrollUserRequest;
 import com.kr.lg.module.user.excpetion.UserException;
 import com.kr.lg.module.user.model.entry.UserAlertEntry;
@@ -28,6 +29,7 @@ public interface UserService {
     void updateUserInfo(UpdateUserInfoRequest request, UserTb userTb) throws UserException, NoSuchAlgorithmException;
     String updateUserProfile(UpdateUserProfileRequest request, UserTb userTb) throws  UserException;
     UserTb enrollUser(EnrollUserRequest request) throws UserException;
+    UserTb enrollUser(LoginDto loginDto) throws UserException;
     void checkOverLapId(String loginId) throws UserException;
     void checkOverLapNickName(String nickName) throws UserException;
 }
